Added security config & changed in the requests and response the balance type to BigDecimal & Implemented the transferAmount.
- I changed it to BigDecimal since float could have some problems in calculations.
- IMP!!!! the transferAmount method is Implemnted BUT since the transaction service is not completed yet it cannot be finished now.

+++ accountService/src/main/java/com/AccountService/AccountService/apis/Resources/InRequest/AccountCreationRequest.java
@@ -1,5 +1,7 @@
 package com.AccountService.AccountService.apis.Resources.InRequest;
 
+import java.math.BigDecimal;
+
 import com.AccountService.AccountService.apis.validation.ValidUUID;
 import com.AccountService.AccountService.apis.validation.ValueOfEnum;
 import com.AccountService.AccountService.applications.enums.AccountType;
@@ -26,5 +28,5 @@ public class AccountCreationRequest {
     @NotNull(message = "initialBalance cannot be empty")
     @NotBlank(message = "initialBalance cannot be empty")
     @PositiveOrZero(message = "Initial balance must be zero or positive")
-    private Float initialBalance;
+    private BigDecimal initialBalance;
 }
