+++ accountService/src/main/java/com/AccountService/AccountService/apis/Resources/OutResponse/AccountDetailResponse.java
@@ -1,5 +1,7 @@
 package com.AccountService.AccountService.apis.Resources.OutResponse;
 
+import java.math.BigDecimal;
+
 import com.AccountService.AccountService.applications.enums.AccountStatus;
 import com.AccountService.AccountService.applications.enums.AccountType;
 import lombok.AllArgsConstructor;
@@ -14,7 +16,7 @@ public class AccountDetailResponse {
 
     private String accountId;
     private String accountNumber;
-    private Float balance;
+    private BigDecimal balance;
     private AccountType accountType;
     private AccountStatus status;
 }
